requirements.txt
anthropic==0.34.2
supabase==2.7.4
python-dotenv==1.0.1
flask==3.0.3
gunicorn==21.2.0
pytz==2025.2
schedule==1.2.2
sendgrid==6.11.0
requests
stripe==7.0.0
resend==0.7.2
